claim_description_normalizer/requirements.txt
streamlit>=1.36.0

langchain>=0.2.10
langchain-openai>=0.1.20
langchain-community>=0.2.10

chromadb>=0.5.5
sentence-transformers>=2.7.0

pydantic>=2.7.4
python-dotenv>=1.0.1
